Done with all the update and delete APIs

+++ src/handlers/user_handler.rs
@@ -2,7 +2,10 @@ use crate::{
     errors::{Error, Result},
     models::{
         dek_model::Dek,
-        user_model::{SignInPayload, UpdateUserPayload, User, UserEmail, UserResponse},
+        user_model::{
+            SignInPayload, ToggleUserActivationStatusPayload, UpdateUserPayload,
+            UpdateUserRolePayload, User, UserEmail, UserResponse,
+        },
     },
     utils::{
         encryption_utils::{create_dek, decrypt_data, encrypt_data},
@@ -280,7 +283,7 @@ pub async fn update_user_handler(
     println!(">> HANDLER: update_user_handler called");
 
     // check if the payload is empty
-    if payload.email.is_empty() || payload.name.is_empty() || payload.role.is_empty() {
+    if payload.email.is_empty() || payload.name.is_empty() {
         return Err(Error::UpdateUserInvalidPayload {
             message: "Invalid payload".to_string(),
         });
@@ -288,15 +291,43 @@ pub async fn update_user_handler(
 
     let db = state.mongo_client.database("test");
     let collection: Collection<User> = db.collection("users");
+    let collection_dek: Collection<Dek> = db.collection("deks");
+
+    let server_kek = env::var("SERVER_KEK").expect("Server Kek must be set.");
+
+    // encrypt the email using kek
+    let encrypted_email_kek = encrypt_data(&payload.email, &server_kek);
+
+    let cursor_dek = collection_dek
+        .find_one(
+            Some(doc! {
+                "email": encrypted_email_kek.clone(),
+            }),
+            None,
+        )
+        .await
+        .unwrap();
+
+    if cursor_dek.is_none() {
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let dek_data = cursor_dek.unwrap();
+
+    // decrypt the dek using the server kek
+    let uid = decrypt_data(&dek_data.uid, &server_kek);
+
+    // find the user in the users collection using the uid
     let cursor = collection
         .update_one(
             doc! {
-                "email": payload.email.clone(),
+                "uid": uid,
             },
             doc! {
                 "$set": {
                     "name": payload.name.clone(),
-                    "role": payload.role.clone(),
                     "updated_at": DateTime::now(),
                 }
             },
@@ -307,6 +338,7 @@ pub async fn update_user_handler(
 
     let modified_count = cursor.modified_count;
 
+    // Return Error if User is not there
     if modified_count == 0 {
         // send back a 404 to
         return Err(Error::UserNotFound {
@@ -316,6 +348,174 @@ pub async fn update_user_handler(
 
     let res = Json(json!({
         "message": "User updated",
+        "user": *payload,
+    }));
+
+    Ok(res)
+}
+
+pub async fn update_user_role_handler(
+    State(state): State<AppState>,
+    payload: Json<UpdateUserRolePayload>,
+) -> Result<Json<Value>> {
+    println!(">> HANDLER: update_user_role_handler called");
+
+    // check if the payload is empty
+    if payload.email.is_empty() || payload.role.is_empty() {
+        return Err(Error::UpdateUserInvalidPayload {
+            message: "Invalid payload".to_string(),
+        });
+    }
+
+    let db = state.mongo_client.database("test");
+    let collection: Collection<User> = db.collection("users");
+    let collection_dek: Collection<Dek> = db.collection("deks");
+
+    let server_kek = env::var("SERVER_KEK").expect("Server Kek must be set.");
+
+    // encrypt the email using kek
+    let encrypted_email_kek = encrypt_data(&payload.email, &server_kek);
+
+    let cursor_dek = collection_dek
+        .find_one(
+            Some(doc! {
+                "email": encrypted_email_kek.clone(),
+            }),
+            None,
+        )
+        .await
+        .unwrap();
+
+    if cursor_dek.is_none() {
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let dek_data = cursor_dek.unwrap();
+
+    // decrypt the dek using the server kek
+    let uid = decrypt_data(&dek_data.uid, &server_kek);
+    let dek = decrypt_data(&dek_data.dek, &server_kek);
+
+    let encrypted_role = encrypt_data(&payload.role, &dek);
+
+    // find the user in the users collection using the uid
+    let cursor = collection
+        .update_one(
+            doc! {
+                "uid": uid,
+            },
+            doc! {
+                "$set": {
+                    "role": encrypted_role,
+                    "updated_at": DateTime::now(),
+                }
+            },
+            None,
+        )
+        .await
+        .unwrap();
+
+    let modified_count = cursor.modified_count;
+
+    // Return Error if User is not there
+    if modified_count == 0 {
+        // send back a 404 to
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let res = Json(json!({
+        "message": "User Role updated",
+        "user": *payload,
+    }));
+
+    Ok(res)
+}
+
+pub async fn toggle_user_activation_status(
+    State(state): State<AppState>,
+    payload: Json<ToggleUserActivationStatusPayload>,
+) -> Result<Json<Value>> {
+    println!(">> HANDLER: update_user_role_handler called");
+
+    match payload.is_active {
+        Some(_) => {
+            if payload.email.is_empty() {
+                return Err(Error::UpdateUserInvalidPayload {
+                    message: "Invalid payload".to_string(),
+                });
+            }
+        }
+        None => {
+            return Err(Error::UpdateUserInvalidPayload {
+                message: "Invalid payload".to_string(),
+            });
+        }
+    }
+
+    let db = state.mongo_client.database("test");
+    let collection: Collection<User> = db.collection("users");
+    let collection_dek: Collection<Dek> = db.collection("deks");
+
+    let server_kek = env::var("SERVER_KEK").expect("Server Kek must be set.");
+
+    // encrypt the email using kek
+    let encrypted_email_kek = encrypt_data(&payload.email, &server_kek);
+
+    let cursor_dek = collection_dek
+        .find_one(
+            Some(doc! {
+                "email": encrypted_email_kek.clone(),
+            }),
+            None,
+        )
+        .await
+        .unwrap();
+
+    if cursor_dek.is_none() {
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let dek_data = cursor_dek.unwrap();
+
+    // decrypt the dek using the server kek
+    let uid = decrypt_data(&dek_data.uid, &server_kek);
+
+    // find the user in the users collection using the uid
+    println!(">> isActive: {:?}", payload.is_active);
+    let cursor = collection
+        .update_one(
+            doc! {
+                "uid": uid,
+            },
+            doc! {
+                "$set": {
+                    "is_active": payload.is_active.unwrap(),
+                    "updated_at": DateTime::now(),
+                }
+            },
+            None,
+        )
+        .await
+        .unwrap();
+
+    let modified_count = cursor.modified_count;
+
+    // Return Error if User is not there
+    if modified_count == 0 {
+        // send back a 404 to
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let res = Json(json!({
+        "message": "User Activation Status updated",
         "user": *payload,
     }));
 
@@ -418,10 +618,38 @@ pub async fn delete_user_handler(
 
     let db = state.mongo_client.database("test");
     let collection: Collection<User> = db.collection("users");
+    let collection_dek: Collection<Dek> = db.collection("deks");
+
+    let server_kek = env::var("SERVER_KEK").expect("Server Kek must be set.");
+
+    // encrypt the email using kek
+    let encrypted_email_kek = encrypt_data(&payload.email, &server_kek);
+
+    let cursor_dek = collection_dek
+        .find_one(
+            Some(doc! {
+                "email": encrypted_email_kek.clone(),
+            }),
+            None,
+        )
+        .await
+        .unwrap();
+
+    if cursor_dek.is_none() {
+        return Err(Error::UserNotFound {
+            message: "User not found".to_string(),
+        });
+    }
+
+    let dek_data = cursor_dek.unwrap();
+
+    // decrypt the uid using the server kek
+    let uid = decrypt_data(&dek_data.uid, &server_kek);
+
     let cursor = collection
         .delete_one(
             doc! {
-                "email": payload.email.clone(),
+                "uid": uid,
             },
             None,
         )
@@ -434,6 +662,26 @@ pub async fn delete_user_handler(
         // send back a 404 to
         return Err(Error::UserNotFound {
             message: "User not found".to_string(),
+        });
+    }
+
+    // delete the dek from the deks collection
+    let cursor_dek_delete = collection_dek
+        .delete_one(
+            doc! {
+                "uid": dek_data.uid,
+            },
+            None,
+        )
+        .await
+        .unwrap();
+
+    let deleted_count_dek = cursor_dek_delete.deleted_count;
+
+    if deleted_count_dek == 0 {
+        // send back a 404 to
+        return Err(Error::UserNotFound {
+            message: "DEK not found".to_string(),
         });
     }
